+++ package.json
@@ -92,7 +92,7 @@
     "moment": "^2.29.1",
     "moment-duration-format": "^2.3.2",
     "moment-timezone": "^0.5.33",
-    "openstack-uicore-foundation": "5.0.17",
+    "openstack-uicore-foundation": "5.0.20-beta.1",
     "p-limit": "^6.1.0",
     "path-browserify": "^1.0.1",
     "postcss-loader": "^6.2.1",
